# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CorruptionFreeTN API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Update a user's role (super_admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRoleInput"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [admin]
      summary: List audit logs (admin only)
      parameters:
        - name: entityType
          in: query
          required: false
          schema:
            type: string
        - name: action
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogsResponse"
        "403":
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/departments:
    get:
      operationId: adminListDepartments
      tags: [admin]
      summary: List all departments (super-admin master data)
      responses:
        "200":
          description: Department list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: adminCreateDepartment
      tags: [admin]
      summary: Create a new department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/departments/{departmentId}:
    patch:
      operationId: adminUpdateDepartment
      tags: [admin]
      summary: Update a department
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: adminDeleteDepartment
      tags: [admin]
      summary: Delete a department
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/districts:
    get:
      operationId: adminListDistricts
      tags: [admin]
      summary: List all districts (super-admin master data)
      responses:
        "200":
          description: District list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminDistrict"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: adminCreateDistrict
      tags: [admin]
      summary: Create a district
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DistrictInput"
      responses:
        "201":
          description: Created district
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDistrict"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/districts/{districtId}:
    patch:
      operationId: adminUpdateDistrict
      tags: [admin]
      summary: Update a district
      parameters:
        - name: districtId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DistrictInput"
      responses:
        "200":
          description: Updated district
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDistrict"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: adminDeleteDistrict
      tags: [admin]
      summary: Delete a district
      parameters:
        - name: districtId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      summary: List all complaint categories (super-admin master data)
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCategory"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Create a complaint category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCategory"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/categories/{categoryId}:
    patch:
      operationId: adminUpdateCategory
      tags: [admin]
      summary: Update a complaint category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCategory"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      summary: Delete a complaint category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/settings:
    get:
      operationId: adminListSettings
      tags: [admin]
      summary: List all portal settings
      responses:
        "200":
          description: Settings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SettingItem"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/settings/{key}:
    patch:
      operationId: adminUpdateSetting
      tags: [admin]
      summary: Update a portal setting value
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingInput"
      responses:
        "200":
          description: Updated setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingItem"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Department:
      type: object
      properties:
        id:
          type: integer
        ministryId:
          type: ["integer", "null"]
        name:
          type: string
        nameTa:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        secretary:
          type: ["string", "null"]
        commissioner:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
      required: [id, name]
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
        departmentId:
          type: ["integer", "null"]
        districtId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, role, createdAt]
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
      required: [users, total]
...
    UpdateUserRoleInput:
      type: object
      properties:
        role:
          type: string
          enum:
            - citizen
            - village_officer
            - taluk_officer
            - district_officer
            - department_officer
            - ministry_officer
            - state_administrator
            - super_admin
            - investigation_officer
            - moderator
            - auditor
            - legal_officer
        departmentId:
          type: ["integer", "null"]
        districtId:
          type: ["integer", "null"]
      required: [role]
...
    AuditLogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        total:
          type: integer
      required: [logs, total]
...
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
        description:
          type: string
    AdminDistrict:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameTa:
          type: ["string", "null"]
        code:
          type: string
      required: [id, name, code]
    DistrictInput:
      type: object
      required: [name, nameTa, code]
      properties:
        name:
          type: string
          minLength: 2
        nameTa:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 2
    AdminCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameTa:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required: [id, name]
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
        nameTa:
          type: string
        description:
          type: string
    SettingItem:
      type: object
      properties:
        key:
          type: string
        value:
          type: ["string", "null"]
        updatedAt:
          type: string
      required: [key]
    SettingInput:
      type: object
      required: [value]
      properties:
        value:
          type: string
...